requirements.txt
# Core dependencies
crewai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.1
python-dotenv>=0.19.0
pyyaml>=6.0.1

# API integrations
openai>=1.0.0
tavily-python>=0.1.0

# Utilities
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0

# Logging and monitoring
logging>=0.5.1.2

# Type checking
typing-extensions>=4.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development tools
black>=22.0.0
isort>=5.0.0
mypy>=0.900
flake8>=4.0.0
pre-commit>=2.17.0